+++ app/(service)/favorite_goods/page.tsx
@@ -1,9 +1,18 @@
+'use client';
+
+import FavoriteProducts from '@/components/FavoriteProducts';
+import Section from '@/components/Section';
 import React from 'react';
 
 type Props = {};
 
 function FavoriteGoodsPage(props: Props) {
-  return <div>좋아요한 상품</div>;
+  return (
+    <div>
+      <Section sectionName="좋아요한 상품" />
+      <FavoriteProducts isPreview={false} />
+    </div>
+  );
 }
 
 export default FavoriteGoodsPage;
